Add DB migrations tests (#899)
* Add migrations test

* Add schemas

* Add woodpecker test

* Put migrations in separate file

* New woodpecker attempt

* Fix formatting

* Remove woodpecker test it requires hardware acceleration

* Add migration testing to contrib

* Trigger woodpecker

* Fix prettier

* Trigger woodpecker

* Fix formatting

---------
