+++ app/.gitignore
@@ -1,3 +1,2 @@
 /build
 /release
-/schemas
