+++ app/src/androidTest/java/com/jerboa/ExampleInstrumentedTest.kt
@@ -2,7 +2,7 @@ package com.jerboa
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import org.junit.Assert.assertEquals
+import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
 
@@ -17,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.jerboa", appContext.packageName)
+        assertTrue(appContext.packageName.startsWith("com.jerboa"))
     }
 }
